+++ .gitignore
@@ -35,6 +35,5 @@ out/
 
 ### VS Code ###
 .vscode/
+!**/resources
 
-# Spring Boot application configuration file
-src/main/resources/application.yml
