+++ src/main/java/com/example/omg_project/domain/user/repository/UserRepository.java
@@ -10,7 +10,6 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
-
     boolean existsByUsernick(String userName);
     boolean existsByUsername(String email);
 }
