+++ src/main/java/com/example/omg_project/global/security/CustomUserDetailsService.java
@@ -1,6 +1,5 @@
 package com.example.omg_project.global.security;
 
-import com.example.omg_project.domain.role.entity.Role;
 import com.example.omg_project.domain.user.entity.User;
 import com.example.omg_project.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,11 +21,9 @@ public class CustomUserDetailsService implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         log.info("CustomUserDetailsService -- 로그인 요청이 들어왔습니다.");
-        log.info("조회할 아이디 : {} ", username);
 
         Optional<User> userOptional = userRepository.findByUsername(username);
         if (userOptional.isEmpty()) {
-            log.info("사용자가 없습니다.");
             throw new UsernameNotFoundException("사용자가 존재하지 않습니다." + username);
         }
