+++ src/main/resources/templates/main/loginform.html
@@ -12,9 +12,6 @@
     <input type="password" name="password" />
     <button>로그인</button>
 </form>
-<a href="/oauth2/authorization/kakao">카카오 로그인</a>
-<a href="/oauth2/authorization/naver">네이버 로그인</a>
-<a href="/oauth2/authorization/google">구글 로그인</a>
-<a href="/api/users/signup">회원가입을 아직 하지 않으셨나요?</a>
+<a href="/signup">회원가입을 아직 하지 않으셨나요?</a>
 </body>
 </html>
