+++ src/main/resources/templates/main/signupform.html
@@ -1,70 +1,20 @@
 <!DOCTYPE html>
-<html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>회원가입 페이지</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            padding: 0;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            background-color: #f4f4f4;
-        }
-
-        h1 {
-            color: #333;
-        }
-
-        form {
-            background-color: #fff;
-            padding: 20px;
-            border-radius: 5px;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-        }
-
-        label {
-            display: block;
-            margin-bottom: 5px;
-            font-weight: bold;
-        }
-
-        input {
-            width: 100%;
-            padding: 10px;
-            margin-bottom: 10px;
-            border: 1px solid #ccc;
-            border-radius: 5px;
-        }
-
-        button {
-            width: 100%;
-            padding: 10px;
-            background-color: #5cb85c;
-            border: none;
-            border-radius: 5px;
-            color: #fff;
-            font-weight: bold;
-            cursor: pointer;
-        }
-
-        button:hover {
-            background-color: #4cae4c;
-        }
-
-        .error {
-            color: red;
-            font-size: 0.9em;
-        }
-    </style>
+    <title>회원가입</title>
+    <link rel="stylesheet" href="/css/signup.css">
 </head>
 <body>
-<form action="/api/users/signup" method="post">
+<form action="/signup" method="post">
     <h1>회원가입</h1>
     <hr/>
+    <div th:if="${error}" class="alert error">
+        <p th:text="${error}"></p>
+    </div>
+    <div th:if="${success}" class="alert success">
+        <p th:text="${success}"></p>
+    </div>
 
     <label for="email">Email</label>
     <input type="email" id="email" name="username" placeholder="Email" required/>
