+++ src/main/resources/templates/user/home.html
@@ -6,8 +6,8 @@
 </head>
 <body>
 <nav>
-    <a href="/api/users/signup">로그인</a>
-    <a href="/api/users/login">회원가입</a>
+    <a href="/api/users/signup">회원가입</a>
+    <a href="/api/users/login">로그인</a>
 </nav>
 <h2>로그인 후 이용해주세요.</h2>
