+++ src/main/resources/templates/user/mypage.html
@@ -6,7 +6,8 @@
 </head>
 <body>
 <h2>로그인에 성공하셨습니다. 축하합니다!</h2>
-<p>안녕하세요, 당신의 이메일 주소 :: <span th:text="${user.username}"></span>님</p>
-<p>안녕하세요, 당신의 이름 :: <span th:text="${user.name}"></span></p>
+<p>당신의 이메일 주소 :: <span th:text="${user.username}"></span>님</p>
+<p>당신의 이름 :: <span th:text="${user.name}"></span></p>
+<p>당신의 생일 :: <span th:text="${user.birthdate}"></span></p>
 </body>
 </html>
